Add styles files to separate styled-components
