+++ src/components/AlarmNavbar/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import AlarmCard from './AlarmCard.jsx';
-import { sortAlarms } from '../util/index';
+import AlarmCard from '../AlarmCard/index.jsx';
+import { sortAlarms } from '../../util/index';
 
 export default function AlarmNavbar({ alarms, onToggleClick, onDeleteButtonClick }) {
   const handleDeleteButtonClick = (event) => onDeleteButtonClick(event.target.id);
