+++ src/components/Spinner/index.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Spinner as S } from './styles';
+
+export default function Spinner() {
+  return (
+    <>
+      <S.StyledSpinner viewBox="0 0 50 50">
+        <circle
+          className="path"
+          cx="25"
+          cy="25"
+          r="20"
+          fill="none"
+          strokeWidth="6"
+        />
+      </S.StyledSpinner>
+    </>
+  );
+}
