+++ src/pages/Login/styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Login = {};
+
+Login.Container = styled.div`
+  height: 100vh;
+`;
