+++ src/pages/Register/styles.js
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const Register = {};
+
+Register.Container = styled.div`
+  height: 100vh;
+`;
